Move tensor utils to common utils
