+++ src/tests/functional/plugin/cpu/single_layer_tests/activation.cpp
@@ -4,7 +4,7 @@
 
 #include <shared_test_classes/single_layer/activation.hpp>
 #include "shared_test_classes/base/ov_subgraph.hpp"
-#include "functional_test_utils/ov_tensor_utils.hpp"
+#include <common_test_utils/ov_tensor_utils.hpp>
 #include "test_utils/cpu_test_utils.hpp"
 
 using namespace InferenceEngine;
