+++ src/tests/functional/plugin/cpu/single_layer_tests/convolution_backprop_data.cpp
@@ -6,7 +6,7 @@
 #include "test_utils/convolution_params.hpp"
 #include "test_utils/fusing_test_utils.hpp"
 #include "shared_test_classes/base/ov_subgraph.hpp"
-#include "functional_test_utils/ov_tensor_utils.hpp"
+#include <common_test_utils/ov_tensor_utils.hpp>
 #include "ngraph_functions/builders.hpp"
 #include <shared_test_classes/single_layer/convolution_backprop_data.hpp>
 #include "openvino/core/preprocess/pre_post_process.hpp"
