+++ src/tests/functional/plugin/cpu/single_layer_tests/ctc_greedy_decoder_seq_len.cpp
@@ -4,7 +4,7 @@
 
 #include <gtest/gtest.h>
 
-#include <functional_test_utils/ov_tensor_utils.hpp>
+#include <common_test_utils/ov_tensor_utils.hpp>
 #include <ngraph_functions/builders.hpp>
 #include <string>
 #include <tuple>
