+++ src/tests/functional/plugin/cpu/single_layer_tests/deformable_convolution.cpp
@@ -5,7 +5,7 @@
 #include "test_utils/cpu_test_utils.hpp"
 #include "ngraph_functions/builders.hpp"
 #include "ngraph_functions/utils/ngraph_helpers.hpp"
-#include <functional_test_utils/ov_tensor_utils.hpp>
+#include <common_test_utils/ov_tensor_utils.hpp>
 #include "shared_test_classes/base/ov_subgraph.hpp"
 
 using namespace InferenceEngine;
