+++ src/tests/functional/plugin/cpu/single_layer_tests/eltwise.cpp
@@ -4,7 +4,7 @@
 
 #include <shared_test_classes/single_layer/eltwise.hpp>
 #include <ngraph_functions/builders.hpp>
-#include "functional_test_utils/ov_tensor_utils.hpp"
+#include <common_test_utils/ov_tensor_utils.hpp>
 #include "test_utils/fusing_test_utils.hpp"
 
 using namespace InferenceEngine;
