+++ src/tests/functional/plugin/cpu/single_layer_tests/lrn.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "functional_test_utils/ov_tensor_utils.hpp"
+#include <common_test_utils/ov_tensor_utils.hpp>
 #include "test_utils/cpu_test_utils.hpp"
 
 #include "shared_test_classes/base/ov_subgraph.hpp"
