+++ src/tests/functional/plugin/cpu/single_layer_tests/nonzero.cpp
@@ -7,7 +7,7 @@
 
 #include "ngraph_functions/builders.hpp"
 #include "ngraph_functions/utils/ngraph_helpers.hpp"
-#include "functional_test_utils/ov_tensor_utils.hpp"
+#include <common_test_utils/ov_tensor_utils.hpp>
 
 using namespace InferenceEngine;
 using namespace CPUTestUtils;
