+++ src/tests/functional/plugin/cpu/single_layer_tests/normalize.cpp
@@ -6,7 +6,7 @@
 #include "test_utils/fusing_test_utils.hpp"
 #include "ngraph_functions/builders.hpp"
 #include "shared_test_classes/base/ov_subgraph.hpp"
-#include "functional_test_utils/ov_tensor_utils.hpp"
+#include <common_test_utils/ov_tensor_utils.hpp>
 
 using namespace ngraph;
 using namespace InferenceEngine;
