+++ src/tests/functional/plugin/cpu/single_layer_tests/one_hot.cpp
@@ -3,7 +3,7 @@
 //
 
 #include <ngraph_functions/builders.hpp>
-#include <functional_test_utils/ov_tensor_utils.hpp>
+#include <common_test_utils/ov_tensor_utils.hpp>
 #include "test_utils/cpu_test_utils.hpp"
 #include "shared_test_classes/base/ov_subgraph.hpp"
