+++ src/tests/functional/plugin/cpu/single_layer_tests/roi_pooling.cpp
@@ -5,7 +5,7 @@
 
 #include "shared_test_classes/base/ov_subgraph.hpp"
 #include "common_test_utils/file_utils.hpp"
-#include "functional_test_utils/ov_tensor_utils.hpp"
+#include <common_test_utils/ov_tensor_utils.hpp>
 #include "common_test_utils/data_utils.hpp"
 
 #include "ie_common.h"
