+++ src/tests/functional/plugin/cpu/single_layer_tests/tensor_iterator.cpp
@@ -5,7 +5,7 @@
 #include <shared_test_classes/single_layer/tensor_iterator.hpp>
 #include "shared_test_classes/base/ov_subgraph.hpp"
 #include "ngraph_functions/builders.hpp"
-#include "functional_test_utils/ov_tensor_utils.hpp"
+#include <common_test_utils/ov_tensor_utils.hpp>
 
 using namespace InferenceEngine;
 using namespace ov;
