+++ src/tests/functional/plugin/cpu/subgraph_tests/src/eltwise_chain.cpp
@@ -10,7 +10,7 @@
 #include <shared_test_classes/base/ov_subgraph.hpp>
 #include <ngraph_functions/builders.hpp>
 #include "common_test_utils/common_utils.hpp"
-#include "functional_test_utils/ov_tensor_utils.hpp"
+#include <common_test_utils/ov_tensor_utils.hpp>
 #include "functional_test_utils/skip_tests_config.hpp"
 #include "test_utils/cpu_test_utils.hpp"
