+++ src/tests/functional/plugin/cpu/subgraph_tests/src/fuse_non0_output_port.cpp
@@ -7,7 +7,7 @@
 #include "ngraph_functions/utils/ngraph_helpers.hpp"
 #include "ngraph_functions/builders.hpp"
 #include "ngraph/runtime/aligned_buffer.hpp"
-#include "functional_test_utils/ov_tensor_utils.hpp"
+#include <common_test_utils/ov_tensor_utils.hpp>
 
 using namespace ov::test;
