+++ src/tests/functional/plugin/cpu/subgraph_tests/src/reshape_fc.cpp
@@ -3,7 +3,7 @@
 //
 
 #include "shared_test_classes/base/ov_subgraph.hpp"
-#include "functional_test_utils/ov_tensor_utils.hpp"
+#include <common_test_utils/ov_tensor_utils.hpp>
 #include "ngraph_functions/builders.hpp"
 #include "test_utils/fusing_test_utils.hpp"
