+++ src/tests/functional/shared_test_classes/src/base/utils/compare_results.cpp
@@ -6,7 +6,7 @@
 
 #include "shared_test_classes/base/layer_test_utils.hpp"
 #include "shared_test_classes/base/utils/compare_results.hpp"
-#include "functional_test_utils/ov_tensor_utils.hpp"
+#include <common_test_utils/ov_tensor_utils.hpp>
 
 namespace ov {
 namespace test {
