+++ src/tests/functional/shared_test_classes/src/single_layer/experimental_detectron_prior_grid_generator.cpp
@@ -5,7 +5,7 @@
 #include "shared_test_classes/single_layer/experimental_detectron_prior_grid_generator.hpp"
 #include "ngraph_functions/builders.hpp"
 #include "common_test_utils/data_utils.hpp"
-#include "functional_test_utils/ov_tensor_utils.hpp"
+#include <common_test_utils/ov_tensor_utils.hpp>
 
 namespace ov {
 namespace test {
