+++ src/tests/functional/shared_test_classes/src/single_layer/matrix_nms.cpp
@@ -3,7 +3,7 @@
 //
 
 #include "ngraph_functions/builders.hpp"
-#include "functional_test_utils/ov_tensor_utils.hpp"
+#include <common_test_utils/ov_tensor_utils.hpp>
 #include "shared_test_classes/single_layer/matrix_nms.hpp"
 #include "shared_test_classes/base/layer_test_utils.hpp"
